requirements.txt
# Dependencias básicas
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# LLM orquestador (OpenAI)
langchain-core>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# Cliente MCP para consumir agentes especializados
langchain-mcp-adapters>=0.1.0
